backend/requirements/base.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
django-redis>=5.4
django-celery-beat>=2.6
django-storages[boto3]>=1.14
channels>=4.1
channels-redis>=4.2
django-ratelimit>=4.1
celery[redis]>=5.4
psycopg[binary]>=3.1
Pillow>=10.0
boto3>=1.34
openpyxl>=3.1
gunicorn>=22.0
uvicorn[standard]>=0.30
whitenoise>=6.7
python-dotenv>=1.0
